File Explorer Toolbar Links Item 1.1.1: Fix injection on current Windows 11 builds (#1783)
Injection broke in a patch update to 26100. 26100 RTM, as well as all prior Windows 11 builds, are just fine.

The method `CShellBrowser::_OnViewMenuPopup`, as well as its sole caller `_OnInitMenuPopup`, were inlined into the sole caller `WndProcBS`. As an easy workaround, I simply hooked the window procedure and checked the arguments in the same way as would result in calling `_OnInitMenuPopup`, which produces the same effect.

No other changes were made for this update.