Add Taskbar Start Button Corner Fix mod (#2872)
* Add Taskbar Start Button Corner Fix mod

This mod fixes an issue on multi-monitor setups with left-aligned start menus where it seems sticky corners, or perhaps something else, occludes the start menu button in the very corner of the screen.

The mod works via a mouse hook installed into this magical root InputSite pane on the taskbar, and checking if a mouse down event happens in the corner of the screen, where the issue is.

* Let UIAutomation clean itself up and lazy load

* Use CreateWindowInBand and ditch the worker thread junk

* simplify ModSettingsChanged

* My brain doesnt work (clean up UIAutomation properly)

* Implement process-aware taskbar wnd grabbing. remove atomic.

* cleanup UIAutomation init, add com init'd check before running CoUninitialize

* Remove timer for UIAutomation uninit, replace with RunFromWindowThread